Update to Firefox v2.1
Updated the HTTP 'GET' request for filter Lists (#116)
Fixed a problem when trying to open the blocking wizard a second time during the same browser session (#113)
Improved the survey processing logic (#111)
Removed unnecessary code for other browsers (#108)
Fixed a display issue of the wizards on certain sites (#103)
Cleaned up and formatted the code (#87)
Fixed an issue with the exclude filters (#97)
Added badge text, showing how many ads were blocked (#95)
Added the extension install date to the ping data (#93)
Fixed a JavaScript exception during browser shutdown (#91)
Added context menu support (#78)
Added an option to hide number the of ads blocked shown in the popup menu (#81)
Added YouTube Channel white-list feature (#84)
Updated the Israel/Hebrew filter list forum link on the 'Report an Ad' page (#86)
Improved the performance of the Add-on when using the Malware filter list.  Also added an option to notify
the user when subscribed to the Malware filter list, and Malware is detected. (#75)
Updated the Japanese filter list URL (#73)
Added a user message that is displayed when the white-list and blacklist wizard can't run on older sites with framesets. (#49)
Added the display of the custom subscription URL to filters page (#31)
Added a 'reload page' option to the white list wizard. (#50)
Added Firefox specific messages to the 'Report an Ad' page. (#64)
Implemented a common debug function to be used when users report an issue or ad to the AdBlock support site (#45)
Fixed a bug in the logic to update the AdBlock icon whenever a tab is updated (#61)
In the 'Report an Ad' feature, the extension will now perform a check for known malware
domains in the list of resources.  If resources from a known malware site are found, the
user is notified, and a link with more information is provided.  If no malware resources
are found, the page continues to work as before. (#32)
Removed the Spanish filter list from the options since it is no longer being supported (#51)

+++ data/adblock_start_firefox.js
@@ -1,10 +1,15 @@
+"use strict";
+
 var elementPurger = {
   onPurgeRequest: function(request) {
-    if (request.command === 'purge-elements' &&
+    if (document &&
+        document.location &&
+        document.location.href &&
+        request.command === 'purge-elements' &&
         request.frameUrl === document.location.href.replace(/#.*$/, ""))
       elementPurger._purgeElements(request);
     else {
-      log("ignoring request: " +    document.location.href.replace(/#.*$/, ""));
+      log("ignoring request ", request);
     }
 
   },
